@@ -64,10 +64,11 @@ struct ReconstructionAlignmentEstimator {
                             const std::vector<Y_t>& images2) const {
     CHECK_GE(images1.size(), 3);
     CHECK_GE(images2.size(), 3);
+    CHECK_EQ(images1.size(), images2.size());
 
     std::vector<Eigen::Vector3d> proj_centers1(images1.size());
     std::vector<Eigen::Vector3d> proj_centers2(images2.size());
-    for (size_t i = 0; i < 3; ++i) {
+    for (size_t i = 0; i < images1.size(); ++i) {
       CHECK_EQ(images1[i]->ImageId(), images2[i]->ImageId());
       proj_centers1[i] = images1[i]->ProjectionCenter();
       proj_centers2[i] = images2[i]->ProjectionCenter();
@@ -194,10 +195,17 @@ SimilarityTransform3::SimilarityTransform3(const double scale,
   transform_.matrix() = matrix;
 }
 
-void SimilarityTransform3::Estimate(const std::vector<Eigen::Vector3d>& src,
+bool SimilarityTransform3::Estimate(const std::vector<Eigen::Vector3d>& src,
                                     const std::vector<Eigen::Vector3d>& dst) {
-  transform_.matrix().topLeftCorner<3, 4>() =
-      SimilarityTransformEstimator<3>().Estimate(src, dst)[0];
+  const auto results = SimilarityTransformEstimator<3>().Estimate(src, dst);
+  if (results.empty()) {
+    return false;
+  }
+
+  CHECK_EQ(results.size(), 1);
+  transform_.matrix().topLeftCorner<3, 4>() = results[0];
+
+  return true;
 }
 
 SimilarityTransform3 SimilarityTransform3::Inverse() const {
