@@ -37,7 +37,8 @@
 
 namespace colmap {
 
-OpenGLContextManager::OpenGLContextManager()
+OpenGLContextManager::OpenGLContextManager(int opengl_major_version,
+                                           int opengl_minor_version)
     : parent_thread_(QThread::currentThread()),
       current_thread_(nullptr),
       make_current_action_(new QAction(this)) {
@@ -46,8 +47,8 @@ OpenGLContextManager::OpenGLContextManager()
 
   QSurfaceFormat format;
   format.setDepthBufferSize(24);
-  format.setMajorVersion(3);
-  format.setMinorVersion(2);
+  format.setMajorVersion(opengl_major_version);
+  format.setMinorVersion(opengl_minor_version);
   format.setSamples(4);
   format.setProfile(QSurfaceFormat::CompatibilityProfile);
   context_.setFormat(format);
@@ -57,13 +58,14 @@ OpenGLContextManager::OpenGLContextManager()
   context_.makeCurrent(&surface_);
   CHECK(context_.isValid()) << "Could not create valid OpenGL context";
 
-  connect(make_current_action_, &QAction::triggered, this,
-          [this]() {
-            CHECK_NOTNULL(current_thread_);
-            context_.doneCurrent();
-            context_.moveToThread(current_thread_);
-          },
-          Qt::BlockingQueuedConnection);
+  connect(
+      make_current_action_, &QAction::triggered, this,
+      [this]() {
+        CHECK_NOTNULL(current_thread_);
+        context_.doneCurrent();
+        context_.moveToThread(current_thread_);
+      },
+      Qt::BlockingQueuedConnection);
 }
 
 void OpenGLContextManager::MakeCurrent() {
