@@ -53,7 +53,8 @@ namespace colmap {
 // in it. The context can then be made current in any other thread.
 class OpenGLContextManager : public QObject {
  public:
-  OpenGLContextManager();
+  OpenGLContextManager(int opengl_major_version = 2,
+                       int opengl_minor_version = 1);
 
   // Make the OpenGL context available by moving it from the thread where it was
   // created to the current thread and making it current.
