@@ -65,6 +65,7 @@ OptionManager::OptionManager(bool add_project_options) {
   vocab_tree_matching.reset(new VocabTreeMatchingOptions());
   spatial_matching.reset(new SpatialMatchingOptions());
   transitive_matching.reset(new TransitiveMatchingOptions());
+  image_pairs_matching.reset(new ImagePairsMatchingOptions());
   bundle_adjustment.reset(new BundleAdjustmentOptions());
   mapper.reset(new IncrementalMapperOptions());
   patch_match_stereo.reset(new mvs::PatchMatchOptions());
@@ -78,7 +79,8 @@ OptionManager::OptionManager(bool add_project_options) {
   desc_->add_options()("help,h", "");
 
   AddRandomOptions();
-
+  AddLogOptions();
+  
   if (add_project_options) {
     desc_->add_options()("project_path", config::value<std::string>());
   }
@@ -177,6 +179,7 @@ void OptionManager::AddAllOptions() {
   AddVocabTreeMatchingOptions();
   AddSpatialMatchingOptions();
   AddTransitiveMatchingOptions();
+  AddImagePairsMatchingOptions();
   AddBundleAdjustmentOptions();
   AddMapperOptions();
   AddPatchMatchStereoOptions();
@@ -419,6 +422,18 @@ void OptionManager::AddTransitiveMatchingOptions() {
                               &transitive_matching->num_iterations);
 }
 
+void OptionManager::AddImagePairsMatchingOptions() {
+  if (added_image_pairs_match_options_) {
+    return;
+  }
+  added_image_pairs_match_options_ = true;
+
+  AddMatchingOptions();
+
+  AddAndRegisterDefaultOption("ImagePairsMatching.block_size",
+                              &image_pairs_matching->block_size);
+}
+
 void OptionManager::AddBundleAdjustmentOptions() {
   if (added_ba_options_) {
     return;
@@ -542,6 +557,8 @@ void OptionManager::AddMapperOptions() {
                               &mapper->mapper.filter_min_tri_angle);
   AddAndRegisterDefaultOption("Mapper.max_reg_trials",
                               &mapper->mapper.max_reg_trials);
+  AddAndRegisterDefaultOption("Mapper.local_ba_min_tri_angle",
+                              &mapper->mapper.local_ba_min_tri_angle);
 
   // IncrementalTriangulator.
   AddAndRegisterDefaultOption("Mapper.tri_max_transitivity",
@@ -730,6 +747,7 @@ void OptionManager::Reset() {
   added_vocab_tree_match_options_ = false;
   added_spatial_match_options_ = false;
   added_transitive_match_options_ = false;
+  added_image_pairs_match_options_ = false;
   added_ba_options_ = false;
   added_mapper_options_ = false;
   added_patch_match_stereo_options_ = false;
@@ -753,6 +771,7 @@ void OptionManager::ResetOptions(const bool reset_paths) {
   *vocab_tree_matching = VocabTreeMatchingOptions();
   *spatial_matching = SpatialMatchingOptions();
   *transitive_matching = TransitiveMatchingOptions();
+  *image_pairs_matching = ImagePairsMatchingOptions();
   *bundle_adjustment = BundleAdjustmentOptions();
   *mapper = IncrementalMapperOptions();
   *patch_match_stereo = mvs::PatchMatchOptions();
@@ -783,6 +802,8 @@ bool OptionManager::Check() {
   if (sequential_matching) success = success && sequential_matching->Check();
   if (vocab_tree_matching) success = success && vocab_tree_matching->Check();
   if (spatial_matching) success = success && spatial_matching->Check();
+  if (transitive_matching) success = success && transitive_matching->Check();
+  if (image_pairs_matching) success = success && image_pairs_matching->Check();
 
   if (bundle_adjustment) success = success && bundle_adjustment->Check();
   if (mapper) success = success && mapper->Check();
